+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to this project will be documented in this file.
 - Insert secrets in a single SQL statement. (#238)
 - Remove unused method in `AuthorizationService`, update `AuthorizationServiceTests`. (#240)
 - Move `ApiKeyRequestFilter` class to `com.iexec.sms.admin` package. (#241)
+- Clean App compute secret endpoints on `AppComputeSecretController`. (#247)
 
 ## [[8.4.0]](https://github.com/iExecBlockchainComputing/iexec-sms/releases/tag/v8.4.0) 2024-01-10